Fix ALF-10684: FTR: Insufficient hooks provided to allow developers to listen to inbound transfers  - TransferListener objects now receive the correct, useful events when files and folders are created, updated, moved, and deleted.  - The processing algorithm has been changed to eliminate (or at least substantially reduce) the risk of a transfer breaking and leaving the filesystem    in a "bad" state.
NOTE: This commit will make any existing FTR instances obsolete due to a schema change on the receiving end. Any FTR instances currently deployed in the community (4.0.a or 4.0.b) will need to be replaced with a clean FTR deployment and repopulated from the source repository - there is no upgrade path.

git-svn-id: http://svn.alfresco.com/repos/alfresco-open-mirror/alfresco/HEAD@32441 7033f55c-6c29-4ad8-afba-765cb616e738