WebDAV: don't show stack traces for ClientAbortException (SocketException) errors.
When mounting WebDAV as a drive on Mac OS X (using finder) many ClientAbortException errors are raised. This appears to be due to the Mac dropping connections whilst the WebDAV handler is still trying to stream content back to the WebDAV client.

Stack traces are now suppressed for DEBUG-level logging, however full stack traces may still be obtained by switching to TRACE for alfresco.webdav.protocol



git-svn-id: http://svn.alfresco.com/repos/alfresco-open-mirror/alfresco/HEAD@35635 7033f55c-6c29-4ad8-afba-765cb616e738